+++ Xplat/src/main/java/vazkii/botania/mixin/BaseSpawnerAccessor.java
@@ -1,0 +1,23 @@
+/*
+ * This class is distributed as part of the Botania Mod.
+ * Get the Source Code in github:
+ * https://github.com/Vazkii/Botania
+ *
+ * Botania is Open Source and distributed under the
+ * Botania License: http://botaniamod.net/license.php
+ */
+
+package vazkii.botania.mixin;
+
+import net.minecraft.util.random.SimpleWeightedRandomList;
+import net.minecraft.world.level.BaseSpawner;
+import net.minecraft.world.level.SpawnData;
+
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(BaseSpawner.class)
+public interface BaseSpawnerAccessor {
+	@Accessor("spawnPotentials")
+	SimpleWeightedRandomList<SpawnData> botania_getSpawnPotentials();
+}